fixing reviews and adding some Aurora dependencys
