+++ app/Entry.js
@@ -15,7 +15,7 @@ import { ExposureHistoryScreen } from './views/ExposureHistory/ExposureHistory';
 import ImportScreen from './views/Import';
 import { LicensesScreen } from './views/Licenses';
 import LocationTracking from './views/LocationTracking';
-import HomeScreen from './views/DR/Home';
+import HomeScreen from './views/DR/HomeScreen/Home.js';
 import Onboarding1 from './views/onboarding/Onboarding1';
 import Onboarding2 from './views/onboarding/Onboarding2';
 import Onboarding3 from './views/onboarding/Onboarding3';
@@ -47,7 +47,7 @@ class Entry extends Component {
     return (
       <NavigationContainer>
         <Stack.Navigator
-          initialRouteName='InitialScreen'
+          initialRouteName='HomeScreen'
           screenOptions={{
             cardStyleInterpolator: CardStyleInterpolators.forHorizontalIOS,
             cardStyle: {
@@ -93,8 +93,8 @@ class Entry extends Component {
             options={{ headerShown: false }}
           />
           <Stack.Screen
-            name='LocationTrackingScreen'
-            component={LocationTracking}
+            name='HomeScreen'
+            component={HomeScreen}
             options={{ headerShown: false }}
           />
           <Stack.Screen
