--- app/views/DR/Home.js
@@ -1,219 +1,0 @@
-import React, { Component } from 'react';
-import {
-  View,
-  TouchableOpacity,
-  Image,
-  StyleSheet,
-  Text,
-  Card,
-} from 'react-native';
-import { Left } from 'native-base';
-import {
-  widthPercentageToDP as wp
-} from 'react-native-responsive-screen';
-import 'moment/locale/es';
-import moment from 'moment';
-//import { Notifications } from 'expo';
-//import { Feels, Aurora } from '../../DR/components/ActionCards'; // <LastBulletin navigation={navigation} />
-import { requestManager } from '../../services/DR/requestManager';
-import styles from '../../components/DR/styles/styles';
-import Colors from '../../constants/DR/colors';
-
-export default class HomeScreen extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      cases: 0,
-      deaths: 0,
-      recovered: 0,
-      todayCases: 0,
-      refreshing: false,
-    };
-  }
-
-  getCases = async () => {
-    try {
-      const {
-        cases,
-        deaths,
-        recovered,
-        todayCases,
-      } = await await requestManager.getAllCases();
-
-      this.setState({
-        cases,
-        deaths,
-        recovered,
-        todayCases,
-        refreshing: false,
-      });
-    } catch (e) {
-      console.log(e);
-    }
-  };
-
-  refresh = () => {
-    this.setState({ refreshing: true });
-    const cases = this.getCases();
-    console.log(cases);
-  };
-
-  async componentDidMount() {
-    const { navigation } = this.props;
-    navigation.setOptions({ headerShown: false });
-    const cases = this.getCases();
-    console.log(cases);
-    this._notificationSubscription = Notifications.addListener(
-      this._handleNotification
-    );
-  }
-
-  getSettings() {
-    return (
-      <TouchableOpacity
-        style={stylesSavePath.settingsContainer}
-        onPress={() => {
-          this.props.navigation.navigate('SettingsScreen');
-        }}>
-      </TouchableOpacity>
-    );
-  }
-
-  _handleNotification = (notification) => {
-    const { navigation } = this.props;
-    const { data } = notification;
-
-    if (data) {
-      const { type } = data;
-      if (type == 'bulletin') {
-        navigation.navigate('App', {
-          screen: 'Noticias',
-          params: {
-            screen: 'NewsNavigation',
-            params: {
-              screen: 'Boletines',
-            },
-          },
-        });
-      }
-    }
-  };
-
-  render() {
-    const date = moment(new Date(), 'DD/MM/YYYY').format('MMMM YYYY');
-    const {
-      props: { navigation },
-      state: { cases, deaths, recovered, todayCases, refreshing },
-    } = this;
-
-    return (
-      <View style={stylesSavePath.settingsContainer}>
-          
-        <View
-          tyle={styles.scrollContainer}
-        >
-          <View style={styles.mainHeader} />
-          <View style={{ margin: wp('2%') }} >
-            <View style={styles.HeaderView}>
-              <View style={styles.rowAndCenter}>
-                <Left>
-                  <Text style={[styles.text, { color: '#fff' }]}>
-                    {date[0].toUpperCase() + date.slice(1)}
-                  </Text>
-                </Left>
-              </View>
-              <Text style={styles.headerText}>COVID-RD</Text>
-            </View>
-
-            <View style={styles.marginAndAlign}>
-              <View style={styles.marginAndAlign}>
-                <View style={styles.actualSituationContent}>
-                  <Text style={[styles.subtitles, { alignSelf: 'center' }]}>
-                    Situación actual
-                  </Text>
-                </View>
-              </View>
-              <View style={styles.actualSituationContainer}>
-                <TouchableOpacity
-                  onPress={() => navigation.navigate('map-cities')}
-                >
-                  <Card style={styles.infoCards}>
-                    <Text style={[styles.dataText]}>{cases}</Text>
-                    <Text style={styles.text}>Positivos</Text>
-                  </Card>
-                </TouchableOpacity>
-                <TouchableOpacity
-                  onPress={() => navigation.navigate('map-cities')}
-                >
-                  <Card style={styles.infoCards}>
-                    <Text
-                      style={[
-                        styles.dataText,
-                        { color: Colors.buttonLightText },
-                      ]}
-                    >
-                      {deaths}
-                    </Text>
-                    <Text style={styles.text}>Fallecidos</Text>
-                  </Card>
-                </TouchableOpacity>
-                <TouchableOpacity
-                  onPress={() => navigation.navigate('map-cities')}
-                >
-                  <Card style={styles.infoCards}>
-                    <Text style={[styles.dataText, { color: Colors.green }]}>
-                      {recovered}
-                    </Text>
-                    <Text style={styles.text}>Recuperados</Text>
-                  </Card>
-                </TouchableOpacity>
-                <TouchableOpacity
-                  onPress={() => navigation.navigate('map-cities')}
-                >
-                  <Card style={styles.infoCards}>
-                    <Text style={[styles.dataText, { color: Colors.orange }]}>
-                      {todayCases}
-                    </Text>
-                    <Text style={styles.text}>Casos del día</Text>
-                  </Card>
-                </TouchableOpacity>
-              </View>
-            </View>
-          </View>
-        </View>
-        {this.getSettings()}
-      </View>
-    );
-  }
-}
-
-const stylesSavePath = StyleSheet.create({
-  backgroundImage: {
-    width: '100%',
-    height: '100%',
-    resizeMode: 'cover',
-    flex: 1,
-    justifyContent: 'flex-end',
-  },
-  mainContainer: {
-    position: 'absolute',
-    flex: 1,
-    // resizeMode: 'contain',
-    // aligns the center of the main container with center of pulse
-    // so that two `flex: 1` views will be have a reasonable chance at natural
-    // flex flow for above and below the pulse.
-    top: '-10%',
-    left: 0.5,
-    right: 0,
-    height: '100%',
-    paddingHorizontal: '12%',
-    paddingBottom: 12,
-  },
-  settingsContainer: {
-    position: 'absolute',
-    top: 0,
-    marginTop: '14%',
-    marginRight: '7%',
-    alignSelf: 'flex-end',
-  },
-});
