+++ app/views/DR/Tabs/News.js
@@ -6,8 +6,8 @@ import {
   StyleSheet,
   View,
 } from 'react-native';
-
-import imgNews from '../../../assets/images/news.jpg';
+//<HeaderImage imgUrl={imgNews} title='News' />
+//import imgNews from '../../../assets/images/news.jpg';
 import Button from '../../../components/Button';
 import HeaderImage from '../../../components/HeaderImage';
 import DataList from '../../../components/List';
@@ -70,7 +70,7 @@ export default function NewsScreen({ navigation }) {
       onBackPress={backToMain.bind(this)}>
       <View style={styles.container}>
         <ScrollView>
-          <HeaderImage imgUrl={imgNews} title='News' />
+          
           <DataList
             data={news}
             navigation={navigation}
