+++ package.json
@@ -63,6 +63,7 @@
     "emotion-theming": "^10.0.27",
     "i18next": "^19.3.3",
     "js-yaml": "^3.13.1",
+    "jsdom-jscore-rn": "0.1.8",
     "lodash": "^4.17.15",
     "native-base": "2.13.8",
     "patch-package": "^6.2.2",
@@ -79,14 +80,13 @@
     "react-native-document-picker": "github:rparet/react-native-document-picker",
     "react-native-fs": "^2.16.6",
     "react-native-gesture-handler": "^1.6.1",
-    "react-native-responsive-screen": "^1.4.0",
     "react-native-linear-gradient": "^2.5.6",
     "react-native-permissions": "^2.0.10",
-    "react-native-vector-icons": "^6.6.0",
     "react-native-popup-menu": "^0.15.7",
     "react-native-pulse": "^1.0.7",
     "react-native-push-notification": "^3.1.9",
     "react-native-reanimated": "^1.7.1",
+    "react-native-responsive-screen": "^1.4.0",
     "react-native-safe-area-context": "0.6.0",
     "react-native-screens": "2.0.0-alpha.12",
     "react-native-share": "^3.1.0",
@@ -94,12 +94,13 @@
     "react-native-splash-screen": "^3.2.0",
     "react-native-svg": "^12.0.3",
     "react-native-uuid-generator": "^6.1.1",
+    "react-native-vector-icons": "^6.6.0",
     "react-native-webview": "^8.1.2",
     "react-native-zip-archive": "^5.0.1",
     "reanimated-bottom-sheet": "^1.0.0-alpha.19",
     "rn-fetch-blob": "^0.12.0",
-    "victory-native": "^34.1.0",
-    "jsdom-jscore-rn": "0.1.8"
+    "utm-latlng": "^1.0.6",
+    "victory-native": "^34.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.8.6",
